# src/docs/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Kafka REST API SwaggerUI",
    "version": "0.1.0",
    "title": "Kafka REST API SwaggerUI",
    "contact": {
      "name": "gnuhpc",
      "url": "https://github.com/gnuhpc",
      "email": "gnuhpc@gmail.com"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "paths": {
...
    "/kafka/brokers": {
      "get": {
        "tags": [
          "kafka-controller"
        ],
        "summary": "List brokers in this cluster",
        "operationId": "listBrokersUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/BrokerInfo"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/kafka/consumer/{topic}/{partition}/{offset}": {
      "get": {
        "tags": [
          "kafka-controller"
        ],
        "summary": "Get the message from the offset of the partition in the topic, decoder is not supported yet",
        "operationId": "getMessageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "topic",
            "in": "path",
            "description": "topic",
            "required": true,
            "type": "string"
          },
          {
            "name": "partition",
            "in": "path",
            "description": "partition",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "offset",
            "in": "path",
            "description": "offset",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "decoder",
            "in": "query",
            "description": "decoder",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/kafka/consumergroup/{consumergroup}": {
      "delete": {
        "tags": [
          "kafka-controller"
        ],
        "summary": "Delete old Consumer Group",
        "operationId": "deleteOldConsumerGroupUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "consumergroup",
            "in": "path",
            "description": "consumergroup",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GeneralResponse"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/kafka/consumergroup/{consumergroup}/{type}/topic/{topic}/lastcommittime": {
      "get": {
        "tags": [
          "kafka-controller"
        ],
        "summary": "getLastCommitTimestamp",
        "operationId": "getLastCommitTimestampUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "consumergroup",
            "in": "path",
            "description": "consumergroup",
            "required": true,
            "type": "string"
          },
          {
            "name": "topic",
            "in": "path",
            "description": "topic",
            "required": true,
            "type": "string"
          },
          {
            "name": "type",
            "in": "path",
            "description": "type",
            "required": true,
            "type": "string",
            "enum": [
              "NEW",
              "OLD"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object",
                "additionalProperties": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/kafka/consumergroup/{consumergroup}/{type}/topic/{topic}/{partition}/{offset}": {
      "put": {
        "tags": [
          "kafka-controller"
        ],
        "summary": "Reset consumer group offset, earliest/latest can be used",
        "operationId": "resetOffsetUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "topic",
            "in": "path",
            "description": "topic",
            "required": true,
            "type": "string"
          },
          {
            "name": "partition",
            "in": "path",
            "description": "partition",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "consumergroup",
            "in": "path",
            "description": "consumergroup",
            "required": true,
            "type": "string"
          },
          {
            "name": "offset",
            "in": "path",
            "description": "offset",
            "required": true,
            "type": "string"
          },
          {
            "name": "type",
            "in": "path",
            "description": "type",
            "required": true,
            "type": "string",
            "enum": [
              "NEW",
              "OLD"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GeneralResponse"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/kafka/consumergroups": {
      "get": {
        "tags": [
          "kafka-controller"
        ],
        "summary": "List all consumer groups from zk and kafka",
        "operationId": "listAllConsumerGroupsUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "type",
            "required": false,
            "type": "string",
            "enum": [
              "NEW",
              "OLD"
            ]
          },
          {
            "name": "topic",
            "in": "query",
            "description": "topic",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/kafka/consumergroups/{consumerGroup}/{type}": {
      "get": {
        "tags": [
          "kafka-controller"
        ],
        "summary": "Describe consumer groups, showing lag and offset, may be slow if multi topic are listened",
        "operationId": "describeCGUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "consumerGroup",
            "in": "path",
            "description": "consumerGroup",
            "required": true,
            "type": "string"
          },
          {
            "name": "type",
            "in": "path",
            "description": "type",
            "required": true,
            "type": "string",
            "enum": [
              "NEW",
              "OLD"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/ConsumerGroupDesc"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/kafka/consumergroups/{consumerGroup}/{type}/topic": {
      "get": {
        "tags": [
          "kafka-controller"
        ],
        "summary": "Get the topics involved of the specify consumer group",
        "operationId": "listTopicByCGUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "consumerGroup",
            "in": "path",
            "description": "consumerGroup",
            "required": true,
            "type": "string"
          },
          {
            "name": "type",
            "in": "path",
            "description": "type",
            "required": true,
            "type": "string",
            "enum": [
              "NEW",
              "OLD"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/kafka/consumergroups/{consumerGroup}/{type}/topic/{topic}": {
      "get": {
        "tags": [
          "kafka-controller"
        ],
        "summary": "Describe consumer groups by topic, showing lag and offset",
        "operationId": "describeCGByTopicUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "consumerGroup",
            "in": "path",
            "description": "consumerGroup",
            "required": true,
            "type": "string"
          },
          {
            "name": "type",
            "in": "path",
            "description": "type",
            "required": true,
            "type": "string",
            "enum": [
              "NEW",
              "OLD"
            ]
          },
          {
            "name": "topic",
            "in": "path",
            "description": "topic",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ConsumerGroupDesc"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/kafka/partitions/add": {
      "post": {
        "tags": [
          "kafka-controller"
        ],
        "summary": "Add a partition to the topic",
        "operationId": "addPartitionUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "addPartition",
            "description": "addPartition",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddPartition"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TopicMeta"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/kafka/partitions/reassign/check": {
      "put": {
        "tags": [
          "kafka-controller"
        ],
        "summary": "Check the partition reassignment process",
        "operationId": "checkReassignPartitionsUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "reassignStr",
            "description": "reassignStr",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "-1": {
            "description": "Reassignment Failed"
          },
          "0": {
            "description": "Reassignment In Progress"
          },
          "1": {
            "description": "Reassignment Completed"
          },
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "integer",
                "format": "int32"
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/kafka/partitions/reassign/execute": {
      "put": {
        "tags": [
          "kafka-controller"
        ],
        "summary": "Execute the partition reassignment",
        "operationId": "executeReassignPartitionsUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "reassignStr",
            "description": "reassignStr",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "integer",
                "format": "int32"
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/kafka/partitions/reassign/generate": {
      "post": {
        "tags": [
          "kafka-controller"
        ],
        "summary": "Generate plan for the partition reassignment",
        "operationId": "generateReassignPartitionsUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "reassignWrapper",
            "description": "reassignWrapper",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ReassignWrapper"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/kafka/topics": {
      "get": {
        "tags": [
          "kafka-controller"
        ],
        "summary": "List topics",
        "operationId": "listTopicsUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "definitions": {
    "AddPartition": {
      "type": "object",
      "properties": {
        "numPartitionsAdded": {
          "type": "integer",
          "format": "int32"
        },
        "replicaAssignment": {
          "type": "string"
        },
        "topic": {
          "type": "string"
        }
      }
    },
    "BrokerInfo": {
      "type": "object",
      "properties": {
        "endPoints": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "host": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "jmxPort": {
          "type": "integer",
          "format": "int32"
        },
        "port": {
          "type": "integer",
          "format": "int32"
        },
        "rack": {
          "type": "string"
        },
        "securityProtocol": {
          "type": "object"
        },
        "startTime": {
          "type": "string",
          "format": "date-time"
        },
        "version": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "GeneralResponse": {
      "type": "object",
      "properties": {
        "msg": {
          "type": "string"
        },
        "state": {
          "type": "string",
          "enum": [
            "success",
            "failure"
          ]
        }
      }
    },
...
    "ReassignWrapper": {
      "type": "object",
      "properties": {
        "brokers": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int32"
          }
        },
        "topics": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
  }
}